+++ crates/hyperswitch_domain_models/src/router_data_v2/flow_common_types.rs
@@ -154,6 +154,9 @@ pub struct RevenueRecoveryRecordBackData;
 pub struct SubscriptionCreateData;
 
 #[derive(Debug, Clone)]
+pub struct CreateCustomerData;
+
+#[derive(Debug, Clone)]
 pub struct UasFlowData {
     pub authenticate_by: String,
     pub source_authentication_id: common_utils::id_type::AuthenticationId,