+++ crates/hyperswitch_domain_models/src/router_request_types/subscriptions.rs
@@ -29,3 +29,22 @@ pub struct SubscriptionCreateRequest {
     pub auto_collection: String,
     pub connector_params: connector_endpoints::ConnectorParams,
 }
+#[derive(Debug, Clone)]
+pub struct CreateCustomerRequest {
+    pub first_name: String,
+    pub last_name: String,
+    pub email: String,
+    pub locale: Option<String>,
+    pub billing_address: Option<BillingAddress>,
+}
+
+#[derive(Debug, Clone)]
+pub struct BillingAddress {
+    pub first_name: String,
+    pub last_name: String,
+    pub line1: String,
+    pub city: String,
+    pub state: String,
+    pub zip: String,
+    pub country: String,
+}