+++ crates/hyperswitch_interfaces/src/api/subscriptions.rs
@@ -2,13 +2,15 @@
 
 use hyperswitch_domain_models::{
     router_flow_types::subscriptions::{
+        CreateCustomer as CreateCustomerFlow, SubscriptionCreate as SubscriptionCreateFlow,
         SubscriptionRecordBack as SubscriptionRecordBackFlow,
-        SubscriptionCreate as SubscriptionCreateFlow,
     },
-    router_request_types::subscriptions::{SubscriptionsRecordBackRequest, SubscriptionCreateRequest},
+    router_request_types::subscriptions::{
+        CreateCustomerRequest, SubscriptionCreateRequest, SubscriptionsRecordBackRequest,
+    },
     router_response_types::{
         revenue_recovery::RevenueRecoveryRecordBackResponse,
-        subscriptions::SubscriptionCreateResponse,
+        subscriptions::{CreateCustomerResponse, SubscriptionCreateResponse},
     },
 };
 
@@ -18,30 +20,30 @@ use super::{ConnectorCommon, ConnectorIntegration};
 /// trait SubscriptionRecordBack for V1
 pub trait SubscriptionRecordBack:
     ConnectorIntegration<
-        SubscriptionRecordBackFlow,
-        SubscriptionsRecordBackRequest,
-        RevenueRecoveryRecordBackResponse,
-    >
+    SubscriptionRecordBackFlow,
+    SubscriptionsRecordBackRequest,
+    RevenueRecoveryRecordBackResponse,
+>
+{
+}
+#[cfg(feature = "v1")]
+/// trait CustomerCreate for V1
+pub trait CreateCustomer:
+    ConnectorIntegration<CreateCustomerFlow, CreateCustomerRequest, CreateCustomerResponse>
 {
 }
 
 #[cfg(feature = "v1")]
 /// trait SubscriptionCreate for V1
 pub trait SubscriptionCreate:
-    ConnectorIntegration<
-        SubscriptionCreateFlow,
-        SubscriptionCreateRequest,
-        SubscriptionCreateResponse,
-    >
+    ConnectorIntegration<SubscriptionCreateFlow, SubscriptionCreateRequest, SubscriptionCreateResponse>
 {
 }
 
-/// trait Subscriptions 
+/// trait Subscriptions
 #[cfg(feature = "v1")]
 pub trait Subscriptions:
-    ConnectorCommon
-    + SubscriptionRecordBack
-    + SubscriptionCreate
+    ConnectorCommon + SubscriptionRecordBack + SubscriptionCreate + CreateCustomer
 {
 }
 
@@ -55,3 +57,7 @@ pub trait Subscriptions {}
 
 #[cfg(not(feature = "v1"))]
 pub trait SubscriptionCreate {}
+
+#[cfg(not(feature = "v1"))]
+/// trait CreateCustomer (disabled when not V1)
+pub trait CreateCustomer {}